Provide support for logrotate facility
Logrotate support is made available by the new additions:

  * public API for closing the log destination file and
  * option for not including time and PID in log filename.

By not including time and PID in the filename, the logging will go to a
single file. If the file exists, it will append to the end. This file
should be set for rotation in logrotate configuration. Clients of glog
should call CloseLogDestination, when a logfile is rotated. The next log
message to be written, will recreate the original file.